requirements.txt
# Core
python-telegram-bot==21.6
python-dotenv==1.0.0

# Real PSTN calls (Twilio)
twilio==9.3.6

# Telegram Voice Chat forwarding
pyrogram==2.0.106
TgCrypto==1.2.5
pytgcalls==0.9.37

# Audio processing
pydub==0.25.1
pyaudio==0.2.14

# Utilities
aiohttp==3.10.5
